refactor(src): move all types to types/index.d.ts
